Changed: Move some files around for namespacing server
