+++ README.md
@@ -394,7 +394,7 @@ See [trace_scoring.rb](./examples/eval/trace_scoring.rb) for a full example.
 
 Run evaluations from the Braintrust web UI against code in your own application.
 
-**Run as a Rack app**
+#### Run as a Rack app
 
 Define evaluators, pass them to the dev server, and start serving:
 
@@ -422,11 +422,38 @@ run Braintrust::Server::Rack.app(
 )
 ```
 
+Add your Rack server to your Gemfile:
+
+```ruby
+gem "rack"
+gem "puma" # recommended
+```
+
+Then start the server:
+
 ```bash
 bundle exec rackup eval_server.ru -p 8300 -o 0.0.0.0
 ```
 
-**Run as a Rails engine**
+See example: [server/eval.ru](./examples/server/eval.ru)
+
+**Custom evaluators**
+
+Evaluators can also be defined as subclasses:
+
+```ruby
+class FoodClassifier < Braintrust::Eval::Evaluator
+  def task
+    ->(input:) { classify(input) }
+  end
+
+  def scorers
+    [Braintrust::Scorer.new("exact_match") { |expected:, output:| output == expected ? 1.0 : 0.0 }]
+  end
+end
+```
+
+#### Run as a Rails engine
 
 Use the Rails engine when your evaluators live inside an existing Rails app and you want to mount the Braintrust eval server into that application.
 
@@ -460,29 +487,16 @@ end
 
 The generator writes `config/initializers/braintrust_server.rb`, where you can review or customize the slug-to-evaluator mapping it discovers from `app/evaluators/**/*.rb` and `evaluators/**/*.rb`.
 
-**Developing locally**
-
-If you want to skip authentication on incoming eval requests while developing locally, set `config.auth = :none` in `config/initializers/braintrust_server.rb`.
-
-That only disables authentication on requests into your Rails app. Any outgoing Braintrust API calls still require normal Braintrust credentials such as `BRAINTRUST_API_KEY`.
-
 See example: [contrib/rails/eval.rb](./examples/contrib/rails/eval.rb)
 
-**Custom evaluators**
+**Developing locally**
 
-Evaluators can also be defined as subclasses:
+If you want to skip authentication on incoming eval requests while developing locally:
 
-```ruby
-class FoodClassifier < Braintrust::Eval::Evaluator
-  def task
-    ->(input:) { classify(input) }
-  end
+- **For Rack**: Pass `auth: :none` to `Braintrust::Server::Rack.app(...)`
+- **For Rails**: Set `config.auth = :none` in `config/initializers/braintrust_server.rb`
 
-  def scorers
-    [Braintrust::Scorer.new("exact_match") { |expected:, output:| output == expected ? 1.0 : 0.0 }]
-  end
-end
-```
+*NOTE: Setting `:none` disables authentication on incoming requests into your server; executing evals requires a `BRAINTRUST_API_KEY` to fetch resources.*
 
 **Supported web servers**
 
@@ -495,14 +509,7 @@ The dev server requires the `rack` gem and a Rack-compatible web server.
 | [Passenger](https://www.phusionpassenger.com/) | 6.x               |                                      |
 | [WEBrick](https://github.com/ruby/webrick)     | Not supported     | Does not support server-sent events. |
 
-Add your chosen server to your Gemfile:
-
-```ruby
-gem "rack"
-gem "puma" # recommended
-```
-
-See examples: [server/eval.ru](./examples/server/eval.ru), [contrib/rails/eval.rb](./examples/contrib/rails/eval.rb)
+See examples: [server/eval.ru](./examples/server/eval.ru), 
 
 ## Documentation
