+++ examples/contrib/rails/eval.rb
@@ -5,7 +5,7 @@
 # This file shows one conventional setup for the Braintrust eval server in Rails:
 #   1. Define evaluator classes under app/evaluators/
 #   2. Generate the initializer with:
-#        bin/rails generate braintrust:eval_server
+#        bin/rails generate braintrust:server
 #   3. Mount the engine in config/routes.rb
 #
 # Requirements:
@@ -31,11 +31,11 @@ end
 # config/initializers/braintrust_server.rb
 # ---------------------------------------------------------------------------
 
-# Generated by: bin/rails generate braintrust:eval_server
+# Generated by: bin/rails generate braintrust:server
 #
-# require "braintrust/server/rails"
+# require "braintrust/contrib/rails/server"
 #
-# Braintrust::Contrib::Rails::Engine.configure do |config|
+# Braintrust::Contrib::Rails::Server::Engine.configure do |config|
 #   config.evaluators = {
 #     "my-classifier" => MyClassifier.new
 #   }
@@ -51,7 +51,7 @@ end
 # ---------------------------------------------------------------------------
 
 # Rails.application.routes.draw do
-#   mount Braintrust::Contrib::Rails::Engine, at: "/braintrust"
+#   mount Braintrust::Contrib::Rails::Server::Engine, at: "/braintrust"
 # end
 
 puts "Braintrust Rails Engine example — see comments for usage"
