--- lib/braintrust/contrib/rails/health_controller.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Braintrust
-  module Contrib
-    module Rails
-      class HealthController < ApplicationController
-        def show
-          render json: {"status" => "ok"}
-        end
-      end
-    end
-  end
-end
