--- lib/braintrust/contrib/rails/list_controller.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module Braintrust
-  module Contrib
-    module Rails
-      class ListController < ApplicationController
-        def show
-          result = Engine.list_service.call
-          render json: result
-        end
-      end
-    end
-  end
-end
