+++ lib/braintrust/contrib/rails/server/routes.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-Braintrust::Contrib::Rails::Engine.routes.draw do
+Braintrust::Contrib::Rails::Server::Engine.routes.draw do
   get "/", to: "health#show"
   get "/list", to: "list#show"
   post "/list", to: "list#show"
